+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # ChangeLog
 
-* Added openUIMask and check if clicked block supports opening UI (Thanks to @Taylor)
-* Bumped up ldlib for better QoL
+* Fixed Tank Widget GUI IO reversed